django/backend/requirements.txt
django
mysql-connector-python
django-cors-headers
